+++ common.go
@@ -1,0 +1,7 @@
+package healthz
+
+// Common check types.
+const (
+	LivenessCheck  = "liveness"
+	ReadinessCheck = "readiness"
+)